--- javascript/index.html
@@ -1,19 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <title>Getting Started</title>
-    <script src="https://unpkg.com/deck.gl/deckgl.min.js"></script>
-    <script src="https://unpkg.com/mapbox-gl/dist/mapbox-gl.js"></script>
-    <link rel="stylesheet" type="text/css" href="https://unpkg.com/mapbox-gl/dist/mapbox-gl.css">
-    <style>
-      #widget {
-        width: 800px;
-        height: 500px;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="widget"></div>
-    <script src="dist/deckgl.js"></script>
-  </body>
-</html>
