+++ javascript/src/index.js
@@ -1,5 +1,14 @@
-import R2DeckGL from "./r2deckgl";
+import DeckGLWidget from "./deckgl-widget";
 
-var r2DeckGL = new R2DeckGL();
-var el = document.getElementById("widget");
-el.innerHTML = r2DeckGL.renderValue();
+HTMLWidgets.widget({
+  name: "deckgl",
+  type: "output",
+  factory: function(el, width, height) {
+    var deckGLWidget = new DeckGLWidget(el, width, height);
+    return {
+      renderValue: deckGLWidget.renderValue,
+      resize: deckGLWidget.resize,
+      getDeckGL: deckGLWidget.getDeckGL
+    };
+  }
+});
