--- javascript/src/r2deckgl.js
@@ -1,7 +1,0 @@
-export default function(el) {
-  console.log("r2deckgl");
-
-  this.renderValue = function() {
-    return "r2deckgl";
-  };
-}
